Printing and Java 1.5 cleanup
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@3928 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/tool/io/GDSLayers.java
@@ -54,7 +54,7 @@ public class GDSLayers
 	public Integer getFirstLayer()
 	{
 		if (normalLayers.size() == 0) return new Integer(0);
-		return (Integer)normalLayers.get(0);
+		return normalLayers.get(0);
 	}
 
 	public int getPinLayer() { return pinLayer; }
